create status free and busy

+++ app/handlers.py
@@ -93,7 +93,7 @@ class status_of_tables():
     async def get_tables(message: Message):
         '''Получение информации о столиках'''
         all_tables = await rq.get_tables()
-        tbl_info = "".join([f"Столик №{t.table_number}: кол-во мест - {t.number_of_seats}, {t.description} - {'Занят❌' if await rq.get_is_occupied_now(t.id) else 'Свободен✅'}\n" for t in all_tables])
+        tbl_info = "".join([f"Столик №{t.table_number}: кол-во мест - {t.number_of_seats}, {t.description} - {'Занят❌' if await rq.get_is_occupied_now(t.table_number) else 'Свободен✅'}\n" for t in all_tables])
         await message.answer(f'Сейчас {datetime.now().strftime("%H:%M")}\n' + tbl_info, reply_markup=kb.tables)
 
     @router.message(F.text == "Изменить")
@@ -115,12 +115,14 @@ class status_of_tables():
     @router.message(and_f(F.text == "Свободен", Form1.status))
     async def status_free(message: Message, state: FSMContext):
         await state.update_data(status=message.text)
+        await rq.set_status_free(int(await state.get_value('table_id')))
         await message.answer(f"Для столика №{await state.get_value('table_id')} задан статус СВОБОДЕН", reply_markup=kb.go_back)
         await state.clear()
 
     @router.message(and_f(F.text == "Занят", Form1.status))
     async def status_busy(message: Message, state: FSMContext):
         await state.update_data(status=message.text)
+        await rq.set_status_busy(int(await state.get_value('table_id')))
         await message.answer(f"Для столика №{await state.get_value('table_id')} задан статус ЗАНЯТ", reply_markup=kb.go_back)
 
 class schedule():
@@ -143,7 +145,7 @@ class schedule():
                                  f"{answer}", reply_markup=kb.booking)
             await state.set_state(Form2.status)
         else:
-            await message.answer("Введите номер столика ещё раз!")
+            await message.answer("Данного столика не существует! Введите номер столика ещё раз!")
 
 class booking():
     '''Здесь прописаны хэндлеры для создания и удаления брони'''
@@ -168,8 +170,11 @@ class booking():
     async def del_booking(message: Message, state: FSMContext):
         await state.update_data(status=message.text)
         delete_time = await rq.get_time_reservation(await state.get_value('table_id'))
-        await message.answer("Выберите время для удаления брони:", reply_markup=kb.go_back)
-        await message.answer("Забронированные часы", reply_markup=kb.delete_time(delete_time))
+        if len(delete_time) > 0:
+            await message.answer("Выберите время для удаления брони:", reply_markup=kb.go_back)
+            await message.answer("Забронированные часы", reply_markup=kb.delete_time(delete_time))
+        else:
+            await message.answer(f"У столика №{await state.get_value('table_id')} нет забронированного времени", reply_markup=kb.booking)
 
     @router.callback_query(F.data.startswith('d_time_'))
     async def conf_del(callback: CallbackQuery, state: FSMContext):
